fix: lint job exiting with 1
In recent versions of Git, a security feature was introduced to prevent directory traversal attacks.
Git now verifies that the directory it’s operating on is “safe.”
If a repository resides in a directory owned by another user
(common in CI environments where repositories are checked out dynamically),
Git might refuse to operate unless explicitly configured.